+++ src/openzeppelin/mod.rs
@@ -1,1 +1,2 @@
 pub mod contracts;
+pub mod contracts_upgradeable;